+++ src/main/java/christmas/service/dto/OrderDto.java
@@ -11,7 +11,7 @@ public class OrderDto {
     private final int totalPrice;
 
     public OrderDto(final Order order) {
-        this.date = order.getDate();
+        this.date = order.getDate().getLocalDate();
         this.menus = order.getMenus();
         this.totalPrice = order.calculateBeforeDiscountTotalPrice();
     }
